requirements.txt
numpy
matplotlib
tensorflow_gpu
Pillow
h5py
pycrypto
gnupg
faced
imageio
matplotlib
cprint
